% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsplitnet.R
\name{geom_splitnet}
\alias{geom_splitnet}
\title{geom_splitnet}
\usage{
geom_splitnet(layout = "slanted", ...)
}
\arguments{
\item{layout}{one of 'rectangular', 'slanted', 'circular', 'radial' or
'unrooted'}

\item{...}{additional parameter}
}
\value{
splitnet layer
}
\description{
add splitnet layer
}
\examples{
data(yeast, package='phangorn')
dm <- phangorn::dist.ml(yeast)
nnet <- phangorn::neighborNet(dm)
ggplot(nnet, aes(x, y))  + geom_splitnet() + theme_tree()
}
\author{
Klaus Schliep
}
