% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsplitnet.R
\name{ggsplitnet}
\alias{ggsplitnet}
\title{ggsplitnet}
\usage{
ggsplitnet(
  tr,
  mapping = NULL,
  layout = "slanted",
  open.angle = 0,
  mrsd = NULL,
  as.Date = FALSE,
  yscale = "none",
  yscale_mapping = NULL,
  ladderize = FALSE,
  right = FALSE,
  branch.length = "branch.length",
  ndigits = NULL,
  ...
)
}
\arguments{
\item{tr}{phylo object}

\item{mapping}{aes mapping}

\item{layout}{one of 'rectangular', 'slanted', 'fan', 'circular', 'radial',
'equal_angle' or 'daylight'}

\item{open.angle}{open angle, only for 'fan' layout}

\item{mrsd}{most recent sampling date}

\item{as.Date}{logical whether using Date class in time tree}

\item{yscale}{y scale}

\item{yscale_mapping}{yscale mapping for category variable}

\item{ladderize}{logical}

\item{right}{logical}

\item{branch.length}{variable for scaling branch, if 'none' draw cladogram}

\item{ndigits}{number of digits to round numerical annotation variable}

\item{...}{additional parameter}
}
\value{
tree
}
\description{
drawing phylogenetic tree from phylo object
}
\examples{
data(yeast, package="phangorn")
dm <- phangorn::dist.ml(yeast)
nnet <- phangorn::neighborNet(dm)
ggsplitnet(nnet) + geom_tiplab2()


}
\references{
Schliep, K., Potts, A. J., Morrison, D. A. and Grimm, G. W.
(2017), Intertwining phylogenetic trees and networks.
\emph{Methods Ecol Evol}. \bold{8}, 1212--1220. doi:10.1111/2041-210X.12760
}
\seealso{
\code{\link[ggtree]{ggtree}}, \code{\link[phangorn]{networx}},
\code{\link[phangorn]{consensusNet}}, \code{\link[phangorn]{neighborNet}}
}
\author{
Klaus Schliep
}
