% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_hybrid_minor.R
\name{swap_hybrid_minor}
\alias{swap_hybrid_minor}
\title{swap_hybrid_minor}
\usage{
swap_hybrid_minor(x, hybrid_nodes, node_times = NULL)
}
\arguments{
\item{x}{evonet object}

\item{hybrid_nodes}{a vector of hybrid nodes to have their minor edges swapped}

\item{node_times}{an optional argument with node times}
}
\value{
network
}
\description{
Swapping the minor edges of an evonet object
}
\examples{
(enet <- ape::read.evonet(text='((a:2,(b:1)#H1:1):1,(#H1,c:1):2);'))
ggevonet(enet) + geom_tiplab()
swapped_enet<-swap_hybrid_minor(enet,6)
ggevonet(swapped_enet) + geom_tiplab()
}
