% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanggle.R
\docType{package}
\name{tanggle-package}
\alias{tanggle}
\alias{tanggle-package}
\title{tanggle: Visualization of Phylogenetic Networks}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Offers functions for plotting split  (or implicit) networks 
    (unrooted, undirected) and explicit networks (rooted, directed) with 
    reticulations extending. 'ggtree' and using functions from 'ape' and 
    'phangorn'. It extends the 'ggtree' package [@Yu2017] to allow the 
    visualization of phylogenetic networks using the 'ggplot2' syntax. It offers
    an alternative to the plot functions already available in 'ape' Paradis and
    Schliep (2019) <doi:10.1093/bioinformatics/bty633> and 'phangorn' Schliep 
    (2011) <doi:10.1093/bioinformatics/btq706>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://klausvigo.github.io/tanggle}
  \item \url{https://github.com/KlausVigo/tanggle}
  \item Report bugs at \url{https://github.com/KlausVigo/tanggle/issues}
}

}
\author{
\strong{Maintainer}: Klaus Schliep \email{klaus.schliep@gmail.com} (\href{https://orcid.org/0000-0003-2941-0161}{ORCID})

Authors:
\itemize{
  \item Marta Vidal-Garcia
  \item Claudia Solis-Lemus \email{solislemus@wisc.edu} (\href{https://orcid.org/0000-0002-9789-8915}{ORCID})
  \item Leann Biancani
  \item Eren Ada
  \item L. Francisco Henao Diaz
}

Other contributors:
\itemize{
  \item Guangchuang Yu [contributor]
}

}
\keyword{internal}
